+++ lib/ui/widgets/account/accounts_widget.dart
@@ -1,4 +1,5 @@
 import 'package:comics_db_app/ui/widgets/account/account_details_cubit.dart';
+import 'package:comics_db_app/ui/widgets/account/account_widget.dart';
 import 'package:comics_db_app/ui/widgets/account/guest_account_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -11,17 +12,12 @@ class AccountsWidget extends StatefulWidget {
 }
 
 class _AccountsWidgetState extends State<AccountsWidget> {
-  // @override
-  // void initState() {
-  //   super.initState();
-  // }
-
   @override
   Widget build(BuildContext context) {
     var cubit = context.watch<AccountDetailsCubit>();
     final String username = cubit.state.username;
     if (username.isNotEmpty) {
-      return const AccountsWidget();
+      return const AccountWidget();
     } else {
       return const GuestAccountWidget();
     }
